requirements.txt
# Core dependencies
flask==2.3.3
flask-cors==4.0.0
sqlalchemy==2.0.20
pydantic==2.3.0
python-dotenv==1.0.0

# API and LLM integration
requests==2.31.0
llama-cpp-python==0.2.11
openai==1.3.0

# Audio processing
SpeechRecognition==3.10.0
pyttsx3==2.90
pyaudio==0.2.13
gTTS==2.3.2
pygame==2.5.0
playsound==1.3.0
py-espeak-ng==0.1.8
azure-cognitiveservices-speech==1.30.0

# Database
sqlite3==0.0.1

# Testing
pytest==7.4.0
pytest-cov==4.1.0

# Utils
numpy==1.25.2
pandas==2.1.0
tqdm==4.66.1 